Move ConfigRevision model & write migrations
