+++ netbox/core/models/__init__.py
@@ -1,3 +1,4 @@
+from .config import *
 from .contenttypes import *
 from .data import *
 from .files import *
