+++ netbox/core/views.py
@@ -1,7 +1,7 @@
 from django.contrib import messages
 from django.shortcuts import get_object_or_404, redirect
 
-from extras.models import ConfigRevision
+from core.models import ConfigRevision
 from netbox.config import get_config
 from netbox.views import generic
 from netbox.views.generic.base import BaseObjectView
